refactor acitivity file to an UI package
