+++ app/src/main/java/com/krisna/marketmaven/ui/ArticleDetailActivity.kt
@@ -1,7 +1,8 @@
-package com.krisna.marketmaven
+package com.krisna.marketmaven.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.krisna.marketmaven.R
 
 class ArticleDetailActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
