+++ app/src/main/java/com/krisna/marketmaven/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.krisna.marketmaven
+package com.krisna.marketmaven.ui
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.krisna.marketmaven.R
 import com.krisna.marketmaven.adapter.ArticleAdapter
 import com.krisna.marketmaven.data.DataArticle
 import com.krisna.marketmaven.data.model.Article
