+++ app/src/main/res/layout/activity_about.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AboutActivity">
+    tools:context=".ui.AboutActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
